[feature] easyapi: complete progress job

+++ annotation/src/main/java/engineer/echo/easyapi/annotation/JobCallback.java
@@ -1,5 +1,5 @@
 package engineer.echo.easyapi.annotation;
 
 public interface JobCallback {
-    void onJobState(State state, int total, int current, int progress);
+    void onJobState(State state, long total, long current, int progress);
 }
